Proposed API implementation
