+++ .gitignore
@@ -6,3 +6,4 @@ test-resources/
 dist/
 *.vsix
 .eslintcache
+vscode.proposed.d.ts
